requirements.txt
Flask==2.1.1
pymongo==4.1.1
dnspython==2.2.1
pyjwt==2.4.0
datetime==4.4